Finish DrawObject refactor
All code for manually drawing is now obsolete. Almost everything in canvas.ts
calls drawDGroup with the name of the group from the manifest.json.

In order to make this work with the window overlay I had to implement a checker
for if part of the window is dirty. This involves some terribly hacky code in
loader.ts; fortunately this only needs to be done once on my machine whenever I
add a new drawObject. The bounds are "baked" meaning they are calculated once
and embedded in the JSON. (I know I'm not really using that term correctly; I'm
too lazy to change it to something better.)

I've done some pretty extensive optimizations and because there's less code to
JIT, it seems the loading is faster. I did end up removing the caching of a
screenshot of the entire canvas in tempData as once it's actually rendering the
VU meters and such at (hopefully) 60 fps, most of the image will need redrawing
anyway. Besides, a full redraw is fast enough now that it's not worth it to keep
the screenshot in memory.

Next order of business is to start migrating the actual hitbox stuff into JSON;
this will be made easier by the bounds rectangles I've implemented for the conf
window.

(I also finally discovered the shortcut for auto-formatting in VS Code, so there
may also be some miscellaneous format improvements.)