Consolidate drawText and drawTextRTL into one function
I added a boolean for if the text should be drawn right-to-left on
the drawText function. I then removed the drawTextRTL function and
all references to it in the DObject code. I changed the manifest so
all text that was previously drawn using textRTL has the boolean
that is passed to drawText. I had to also put spaceWidth after the
rtl boolean as rtl is used much more often than spaceWidth and JSON
doesn't support serializing undefined objects.