Revamp HitDetector and automatically generate HitRegions
HitDetector (in hitbox.ts) now has lists for multiple onMouseDown and
onMouseUp functions. These will allow multiple event handlers at once.
Almost every HitRegion in canvas.ts is now generated automatically by
autoHitRegion, which optionally adds handlers for automatically redrawing
on mouseEnter, mouseLeave, mouseDown, and mouseUp. This simplifies the
code for initializing the hitboxes, and I hope to move all of that into
ui.ts eventually, with the only functions left in canvas.ts being the
raw drawing functions.

Performance is still a bit of an issue, with only ~45 fps. I tried not
changing the ctx.fillStyle or ctx.strokeStyle if it was setting it to
what it already had, but the checks ended up taking more time anyway (I
think it does that check internally anyway).