Render the config window in a separate canvas
This commit moves all of the non-drawing-related functions and variables from
CanvasRenderer into ui.ts, with a new JSSCC class handling all the UI logic.
The config window canvas is overlaid on top of the other, offloading the hard
work of compositing through the transparent overlay to the browser instead of
my own stupid code.

Because the browser is doing the compositing of the overlay, the performance is
vastly increased when the config window is occluding other DrawObjects. The perf
is about equal whether the window is open or closed at this point. I was able to
remove all the hacky code for resetting and redrawing the overlay over an area when
a DrawGroup needed to be redrawn with the config window. Theoretically this version
uses more RAM but considering all the GC pressure the old version had, this approach
is better, all things considered.

I had to refactor a lot of code in AssetLoader as well so it could handle the
duplicate manifests correctly. I now have them isolated so the config window layer
doesn't see the drawObjects of the main UI window and vice versa.

I also moved all of the hitbox initialization code to the new JSSCC class as well.
I hope to get canvas.ts even smaller in the future as I refactor more into the new
ui.ts class.